widget: move GenPollUrl to its own file
The lazy_import() machinery catches all ImportErrors and logs an error if
people try to use a particular widget but don't have the dependencies. As a
result, we don't need this try/catch ImportError for aiohttp.

However, since there are several widgets that live in this file, and all of
them will fail to load if aiohttp is not installed, it's best to move
GenPollUrl to its own file, so that only GenPollUrl is unavailable if
aiohttp is missing.

Fixes #5785
