+++ libqtile/widget/crypto_ticker.py
@@ -2,7 +2,7 @@ import locale
 
 from libqtile.confreader import ConfigError
 from libqtile.log_utils import logger
-from libqtile.widget.generic_poll_text import GenPollUrl
+from libqtile.widget.gen_poll_url import GenPollUrl
 
 _DEFAULT_CURRENCY = str(locale.localeconv()["int_curr_symbol"])
 _DEFAULT_SYMBOL = str(locale.localeconv()["currency_symbol"])
