+++ libqtile/widget/idlerpg.py
@@ -1,6 +1,6 @@
 import datetime
 
-from libqtile.widget.generic_poll_text import GenPollUrl
+from libqtile.widget.gen_poll_url import GenPollUrl
 
 
 class IdleRPG(GenPollUrl):
